Update transformWriter.cpp
A static _AnimChannel is added if the Maya plug is connected, even if the value is the null value. The purpose of this change is to allow USD connections to mimic Maya connections (with a custom primWriter and export chaser) and the USD attribute needs to exist to create the connection with the UsdAttribute::AddConnection function.